+++ NEWS.md
@@ -16,6 +16,8 @@
 
 * Improved width calculations in `as_gtable()` (@teunbrand, #1923)
 
+* Interactive tables now respect more styling options, namely: `table.font.weight`, `stub_background.color`. (#1814)
+
 # gt 0.11.1
 
 ## Breaking changes
@@ -38,7 +40,7 @@
 
 ## Interactive table support
 
-* Interactive tables will show no border if `opt_table_lines(extent = "none")` is specified (#1307).
+* Interactive tables respect`opt_table_lines(extent = "none")` and `opt_table_lines(extent = "all")` is specified (#1307).
 
 * Interactive tables now respect more styling options, namely: `column_labels.background.color`, `row_group.background.color`, `row_group.font.weight`, `table_body.hlines.style`, `table.font.weight`, `table.font.size`, and `stub.font.weight`. (#1693)